(fix) Correct behaviour for quotes handling and tokenization (plus files organization)
