+++ Makefile
@@ -11,10 +11,11 @@ INCLUDE = -I./includes
 
 SD = src/
 COMMON = $(addprefix $(SD), lexer.c lexer_utils.c list_handler.c \
-						parser.c parser_utils.c create_pipeline.c moses_i.c \
+						parser.c parser_utils.c pipeline.c moses_i.c \
 						moses_ii.c hashtable.c expander.c path_finder.c \
-						finisher.c executor.c executor_utils.c \
-						exec_simple_command.c exec_pipeline_commands.c)
+						finisher.c executor.c executor_utils.c conditionals.c \
+						set_structures.c exec_simple_command.c \
+						exec_pipeline_commands.c error_handler.c)
 COMMON += $(addprefix $(SD)builtin/, pwd.c env.c echo.c cd.c exit.c \
 						export.c export_utils.c unset.c)
 SRC = $(COMMON) $(SD)minishell.c
