+++ src/builtin/echo.c
@@ -6,13 +6,13 @@
 /*   By: Juliany Bernardo <julberna@student.42sp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/12 18:47:27 by Juliany Ber       #+#    #+#             */
-/*   Updated: 2024/02/23 16:52:44 by Juliany Ber      ###   ########.fr       */
+/*   Updated: 2024/02/24 18:56:04 by Juliany Ber      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-int	echo_err(void);
+int	echo_error(void);
 int	check_flag(t_word *wl);
 
 int	echo(t_meta *meta, t_word *wl)
@@ -30,19 +30,19 @@ int	echo(t_meta *meta, t_word *wl)
 	{
 		ret = write(STDOUT_FILENO, temp->word, ft_strlen(temp->word));
 		if (ret == -1)
-			return (echo_err());
+			return (echo_error());
 		temp = temp->next;
 		if (temp)
 			ret = write(STDOUT_FILENO, " ", 1);
 		if (ret == -1)
-			return (echo_err());
+			return (echo_error());
 	}
 	if (!flag)
 		write(STDOUT_FILENO, "\n", 1);
 	return (EXIT_SUCCESS);
 }
 
-int	echo_err(void)
+int	echo_error(void)
 {
 	perror("minishell: echo");
 	return (-1);
@@ -50,7 +50,7 @@ int	echo_err(void)
 
 int	check_flag(t_word *wl)
 {
-	if (!ft_strcmp(wl->next->word, "-n"))
+	if (wl->next && !ft_strcmp(wl->next->word, "-n"))
 		return (TRUTH);
 	return (LIE);
 }
