+++ src/list_handler.c
@@ -6,13 +6,11 @@
 /*   By: Juliany Bernardo <julberna@student.42sp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/10 13:48:46 by julberna          #+#    #+#             */
-/*   Updated: 2024/02/13 21:37:33 by Juliany Ber      ###   ########.fr       */
+/*   Updated: 2024/02/24 16:53:34 by Juliany Ber      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/minishell.h"
-
-t_token	*tk_last(t_token *tk);
 
 void	new_token(t_token **tk, int type, char *literal)
 {
